refactor(studio): 收拢知识工作台路由
